fix(glob-routes): split server/client entries
