+++ packages/vite-glob-routes/src/react-router/client.ts
@@ -1,0 +1,18 @@
+import virtualPageRoutesClient from "virtual:@hiogawa/vite-glob-routes/internal/pageRoutesClient";
+import vurtialPageRoutesClientLazy from "virtual:@hiogawa/vite-glob-routes/internal/pageRoutesClientLazy";
+import { createGlobPageRoutes } from "./route-utils";
+
+export function globPageRoutesClient() {
+  return createGlobPageRoutes(virtualPageRoutesClient);
+}
+
+export function globPageRoutesClientLazy() {
+  return createGlobPageRoutes(vurtialPageRoutesClientLazy);
+}
+
+export {
+  initializeClientRoutes,
+  getClientGlobal,
+  getPagePrefetchLinks,
+  type ClientGlobal,
+} from "./cilent-helper";
