+++ packages/vite-glob-routes/src/react-router/route-utils.ts
