+++ packages/vite-glob-routes/src/react-router/server.ts
@@ -1,0 +1,11 @@
+import virtualPageRoutesServer from "virtual:@hiogawa/vite-glob-routes/internal/pageRoutesServer";
+import { createGlobPageRoutes } from "./route-utils";
+
+export function globPageRoutesServer() {
+  return createGlobPageRoutes(virtualPageRoutesServer);
+}
+
+export {
+  handleReactRouterServer,
+  type ServerRouterResult,
+} from "./server-helper";
