+++ packages/vite-glob-routes/tsup.config.ts
@@ -3,7 +3,13 @@ import { defineConfig } from "tsup";
 export default defineConfig({
   // TODO: clean only on release build
   // clean: true,
-  entry: ["src/index.ts", "src/react-router.ts", "src/hattip.ts"],
+  entry: [
+    "src/index.ts",
+    "src/react-router/client.ts",
+    "src/react-router/server.ts",
+    "src/hattip.ts",
+  ],
+  // TODO: probably no point in providing cjs build
   format: ["esm", "cjs"],
   dts: true,
   external: [
